python-lib/requirements.txt
asyncio
pyyaml
psycopg2-binary
ph4_walkingpad
Flask[async]
websockets
pyinstaller
cx_Freeze
